Refactor #114: Tighten cmd/ucm/generate package after code review
Six follow-ups to the C.3b generate-subgroup port (PR #114):

- I1: Export TagsStyleKeys from ucm/config/generate and reference it from
  cmd/ucm/generate/util.go so both packages share one style map. Drops a
  4-line duplicated map.
- I3: writeResourceYAML now os.MkdirAll's the resolved output dir before
  handing it to yamlsaver, mirroring cmd/bundle/generate. Adds a test for
  the previously-broken nested --output-dir path.
- M2: Move stripPreRun and runSubcmd from catalog_test.go into a dedicated
  helpers_test.go (matches cmd/ucm/helpers_test.go convention; deleting any
  one *_test.go file no longer breaks the others).
- M3: Move copyMap from catalog.go into util.go so connection.go's use of
  it isn't co-located via a sibling file.
- M9: Volume Long help now mentions that managed volumes have their
  server-assigned storage_location dropped on emit (UC re-derives on
  deploy).
- M10: Doc comments on every exported NewGenerate*Command factory and the
  parent New() per .agent/rules/style-guide-go.md.

M8 (delete dead ucm/phases/generate.go) deferred to issue #115; it's a
larger cleanup that wants its own PR.
