+++ cmd/ucm/generate/catalog_test.go
@@ -1,59 +1,16 @@
 package generate
 
 import (
-	"bytes"
-	"context"
 	"os"
 	"path/filepath"
 	"testing"
 
-	"github.com/databricks/cli/libs/cmdctx"
-	"github.com/databricks/cli/libs/cmdio"
-	"github.com/databricks/cli/libs/flags"
 	"github.com/databricks/databricks-sdk-go/experimental/mocks"
 	"github.com/databricks/databricks-sdk-go/service/catalog"
-	"github.com/spf13/cobra"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 )
-
-// stripPreRun clears auth-hook PreRunE/PersistentPreRunE recursively. The
-// real generate subcommands set root.MustWorkspaceClient as PreRunE; tests
-// stand in their own mock workspace client via cmdctx so the real hook
-// would just fail on a missing ~/.databrickscfg.
-func stripPreRun(cmd *cobra.Command) {
-	cmd.PersistentPreRunE = nil
-	cmd.PersistentPreRun = nil
-	cmd.PreRunE = nil
-	cmd.PreRun = nil
-	for _, sub := range cmd.Commands() {
-		stripPreRun(sub)
-	}
-}
-
-// runSubcmd executes the named subcommand from the parent generate group
-// with cmdctx pre-populated with the mock workspace client. Returns stderr
-// (captured cmdio output) and the cobra error. Args are passed verbatim;
-// callers include both the subcommand name and any flags.
-func runSubcmd(t *testing.T, w *mocks.MockWorkspaceClient, args ...string) (string, error) {
-	t.Helper()
-	cmd := New()
-	stripPreRun(cmd)
-
-	ctx := context.Background()
-	var stderr bytes.Buffer
-	cmdIO := cmdio.NewIO(ctx, flags.OutputText, nil, &bytes.Buffer{}, &stderr, "", "")
-	ctx = cmdio.InContext(ctx, cmdIO)
-	ctx = cmdctx.SetWorkspaceClient(ctx, w.WorkspaceClient)
-	cmd.SetContext(ctx)
-
-	cmd.SetOut(&bytes.Buffer{})
-	cmd.SetErr(&bytes.Buffer{})
-	cmd.SetArgs(args)
-	err := cmd.Execute()
-	return stderr.String(), err
-}
 
 func TestNewGenerateCatalogCommand_Help(t *testing.T) {
 	cmd := NewGenerateCatalogCommand()
@@ -126,6 +83,26 @@ func TestGenerateCatalog_RefusesOverwriteWithoutForce(t *testing.T) {
 	)
 	require.Error(t, err)
 	assert.Contains(t, err.Error(), "already exists")
+}
+
+func TestGenerateCatalog_CreatesMissingOutputDir(t *testing.T) {
+	work := t.TempDir()
+	nested := filepath.Join(work, "does", "not", "exist")
+
+	w := mocks.NewMockWorkspaceClient(t)
+	w.GetMockCatalogsAPI().EXPECT().
+		GetByName(mock.Anything, "team_alpha").
+		Return(&catalog.CatalogInfo{Name: "team_alpha"}, nil)
+
+	_, err := runSubcmd(t, w,
+		"catalog",
+		"--existing-catalog-name", "team_alpha",
+		"--output-dir", nested,
+	)
+	require.NoError(t, err)
+
+	_, err = os.Stat(filepath.Join(nested, "catalogs_team_alpha.yml"))
+	require.NoError(t, err)
 }
 
 func TestGenerateCatalog_ForceOverwrites(t *testing.T) {
